### 2024-10-09 #7337 @editorialbot
[PRE REVIEW]: Emiproc: A Python package for emission inventory processing
**Submitting author:** @lionel42 (<a href="http://orcid.org/0009-0009-0347-4897">Lionel Constantin</a>)
**Repository:** https://github.com/C2SM-RCM/emiproc
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.1
