### 2024-10-09 #7337 @arfon
> 🔴 License found: Creative Commons Attribution 4.0 International (Not [OSI approved](https://opensource.org/licenses))

Just noting that this seems to be an error on our part – my checking of the repository shows a BSD license so I think there's nothing to do here.
