### 2024-10-10 #7337 @lionel42
> **License info**:
>
> 🔴 License found: `Creative Commons Attribution 4.0 International` (Not [OSI approved](https://opensource.org/licenses))

This is not correct. We have changed the licence to a BSD-3 licence for the new package update
See
https://github.com/C2SM-RCM/emiproc/blob/master/LICENSE.txt
