### 2024-10-23 #7337 @kthyng
@lionel42
* Can this be released to PyPI?
* Please reduce to length of your paper to up to about 1000 words (it is about 1770 words currently). Paper format information here: https://joss.readthedocs.io/en/latest/paper.html
* In your paper, please check that your inline vs. parenthetical references are correct. You can compare with the example paper here: https://joss.readthedocs.io/en/latest/example_paper.html
