### 2024-10-24 b032c9b
Reducing size of paper: removing detailed description
