+++ joss/paper.md
@@ -58,9 +58,6 @@ Inventories are created in multiple different formats and resolutions, which mak
 The package provides functions for performing various 
 operations on inventory data, such as remapping to different model grids, aggregating emissions by sector or pollutant, or scaling emissions based on future scenarios. 
 
-Emission input files can be generated in regular (e.g. hourly) intervals by applying sector- and country-specific temporal and vertical emission profiles. Alternatively, a small set of input files can be generated, which describe the sectorial gridded emissions and their temporal and vertical profiles.
-This set of files can then be read by an OEM (online emissions module) described in [@gmd-13-2379-2020], which applies the temporal and vertical scaling online during the model simulation. OEM is integrated into the atmospheric chemistry and transport models COSMO-ART and ICON-ART.
-
 `emiproc` is designed to be flexible and extendable, allowing users to easily add custom functionality, to read new inventories or export data to other formats.
 
 # Statement of need
@@ -109,96 +106,40 @@ necessary to harmonize them. This is what the `emiproc` package is designed for.
 
 ![Design emiproc. \label{fig:design}](pipeline.drawio.png){ width=60% }
 
-Thanks to the harmonization of the data, functions for additional data processing
-can easily be applied to the different inventories once loaded into `emiproc`. 
-The processed inventories can be exported to any of the formats supported by the package.
+
+
 
 The API of `emiproc` leverages the advantages of object-oriented programming. 
+The main class is `Inventory`. It provides a common data structure
+to represent emission inventories.
+The `Grid` object represent the spatial distribution of the emission.
+It is flexible enough to implement all kinds of grids, from regular latitude-longitude grids
+to the icosahedral grid  from the ICON Model [@IconRelease01].
+Finally `TemporalProfile` / `VerticalProfile` are used for the distribution of emissions 
+in time and height.
+
+Thanks to the harmonization of the data, functions for additional data processing
+can easily be applied to the different inventories once loaded into `emiproc`. 
+These functions are listed in the [API documentation](https://emiproc.readthedocs.io/en/master/api/index.html) 
+For example `remap_inventory()` set an inventory to a different grid.
+
+
+`emiproc` also allows for generation of additional emission sectors not always present in the inventories.
+For example, the vegetation emissions can be modelled through VPRM (Vegetation Photosynthesis and Respiration Model) [@vprm].
+A module for human respiration and for heating emissions have also been implemented.
+
+The export of inventories for various
+atmospheric models is done through custom functions that producs all emission input files required by the
+model. 
+
+For data vizualisation `emiproc` provides custom functions based on `matplotlib` [@matplotlib].
+Figure \ref{fig:tno} was created using this function.
+
 `emiproc` is built on top of the `geopandas` package [@kelsey_jordahl_2020_3946761], 
 which allows storing the geometries of the emission maps and offers many functionalities
 related to geometric operations.
 Within emiproc, the emission data of the inventory is stored as a
 [`geopandas.GeoDataFrame`](https://geopandas.org/en/stable/docs/reference/geodataframe.html).
-
-
-## Inventory
-The main object is the `Inventory` object. Any inventory is a subclass of the base `Inventory`.
-`emiproc` provides many utility functions to process them. 
-Typically, a processing function takes an inventory and some parameters as input 
-and returns a new inventory with the processed data.
-
-Example: the `group_categories` function regroups an inventory with a given set of source
-categories to a new inventory with another set of categories, based on a mapping provided by the user. 
-This is useful to reduce the number of sources simulated or to use a standardized set of 
-categories such as the GNFR (Gridded Nomenclature For Reporting) sectors from [@emep_guidelines].
-
-
-## Grid
[…]
