+++ .github/workflows/publish-pypi.yaml
@@ -20,6 +20,7 @@ jobs:
         python3 -m
         pip install
         build
+        setuptools-scm
         --user
     - name: Build a binary wheel and a source tarball
       run: python3 -m build
