+++ pyproject.toml
@@ -49,6 +49,8 @@ testpaths = ["tests"]
 [tool.setuptools.packages.find]
 include = ["emiproc"]  # package names should match these glob patterns (["*"] by default)
 
+[tool.setuptools_scm]
+write_to = "your_package_name/_version.py"
 
 
 [project.optional-dependencies]
