+++ pyproject.toml
@@ -50,7 +50,7 @@ testpaths = ["tests"]
 include = ["emiproc"]  # package names should match these glob patterns (["*"] by default)
 
 [tool.setuptools_scm]
-write_to = "your_package_name/_version.py"
+write_to = "emiproc/_version.py"
 
 
 [project.optional-dependencies]
