### 2024-10-24 abba3eb
Automatic versionning from tag release for pypi (#67)
* remove github release

* update for version to work
