### 2024-10-24 #7337 @lionel42
@kthyng  Thanks for the feedback

>     * Can this be released to PyPI?
Yes, I just setup the github action to release to pypi: https://pypi.org/project/emiproc/#description
>
>     * Please reduce to length of your paper to up to about 1000 words (it is about 1770 words currently). Paper format information here: https://joss.readthedocs.io/en/latest/paper.html
>
I have reduced the size of some sections. Is there a possibility to trigger the bot again to count the words now ?

>     * In your paper, please check that your inline vs. parenthetical references are correct. You can compare with the example paper here: https://joss.readthedocs.io/en/latest/example_paper.html

I am not sure I understand this point, the references look okay for me.
