+++ docs/source/installation.rst
@@ -1,18 +1,31 @@
 Installation
 ============
 
+With pip
+--------
 
-Installing emiproc can be done as a standard python package.
+The easiest way to install emiproc is to use pip. This will install the latest
+release from PyPI.
+
+.. code:: bash
+
+    pip install emiproc
+
+From source
+-----------
+
 
 First create a folder for emiproc and go into it 
 
-.. code::
+.. code:: bash
 
     mkdir emiproc 
     cd emiproc
 
 
-Clone the repository::
+Clone the repository
+
+.. code:: bash
 
     git clone https://github.com/C2SM-RCM/emiproc.git .
 
@@ -21,7 +34,7 @@ Then install using pip.
 Installing as an editable (`-e` options) allows you to directly modify 
 the cloned code.
 
-.. code::
+.. code:: bash
 
     pip install -e .
 
