### 2024-10-24 #7337 @kthyng
> Yes, I just setup the github action to release to pypi: https://pypi.org/project/emiproc/#description

Great! I see it there. Please update your installation instructions too.
