### 2024-10-24 #7337 @kthyng
> I am not sure I understand this point, the references look okay for me

An an example, line 27 has the reference in parentheses but it shouldn't be.
