### 2024-10-24 #7337 @lionel42
> Great! I see it there. Please update your installation instructions too.

Here it is: installation with pip:

https://emiproc.readthedocs.io/en/master/installation.html
