### 2024-10-24 a2744cc
Update README.md with installation
