### 2024-10-24 c73688e
update with new swiss and added out resolution
