+++ scripts/zh_2_footprintsgrid.py
@@ -24,13 +24,14 @@ from emiproc.inventories.swiss import SwissRasters
 from emiproc.inventories.utils import (
     add_inventories,
     crop_with_shape,
+    drop,
     get_total_emissions,
     scale_inventory,
     group_categories,
 )
 from emiproc.inventories.zurich import MapLuftZurich
 from emiproc.regrid import remap_inventory
-from emiproc.speciation import speciate, speciate_inventory
+from emiproc.speciation import merge_substances, speciate, speciate_inventory
 from emiproc.utilities import SEC_PER_YR
 from emiproc.exports.rasters import export_raster_netcdf
 from emiproc.exports.netcdf import nc_cf_attributes
@@ -50,7 +51,7 @@ mapluft_dir = Path(r"C:\Users\coli\Documents\Data\mapluft_emissionnen_kanton")
 mapluf_file = mapluft_dir / f"mapLuft_{YEAR}_v2024.gdb"
 
 # edge of the raster cells
-VERSION = "v2"
+VERSION = "v2_forBetty"
 
 # Whether to split the biogenic CO2 and the antoropogenic CO2
 SPLIT_BIOGENIC_CO2 = True
@@ -72,6 +73,7 @@ SPLIT_GNRF_ROAD_TRANSPORT = True
 
 # %% Check some parameters and create the output directory
 weights_dir = footprints_dir / f"weights_files_{YEAR}_{VERSION}"
+weights_dir.mkdir(exist_ok=True)
 
 if SPLIT_GNRF_ROAD_TRANSPORT and not USE_GNRF:
     raise ValueError("Cannot split GNRF if not using GNRF")
@@ -130,13 +132,14 @@ ds = ds.assign_coords(datetime=("timestep", datetime))
 x_coords = ds.x_lv95.values
 y_coords = ds.y_lv95.values
 d2 = 5.0  # meters, half the size of the cells
+d_out = 100  # meters, size of the cells of the output grid
 grid = RegularGrid(
     xmin=x_coords.min() - d2,
     xmax=x_coords.max() + d2,
     ymin=y_coords.min() - d2,
     ymax=y_coords.max() + d2,
-    nx=x_coords.size,
-    ny=y_coords.size,
+    dx=d_out,
+    dy=d_out,
     crs="LV95",
 )
 
@@ -147,14 +150,6 @@ if SPLIT_BIOGENIC_CO2:
     from emiproc.inventories.zurich.speciation_co2_bio import ZH_CO2_BIO_RATIOS
 
     inv = speciate(inv, "CO2", ZH_CO2_BIO_RATIOS, drop=True)
-
-# %% do the actual remapping of zurich to rasters
-
-rasters_inv = remap_inventory(
-    crop_with_shape(inv, zh_shape),
-    grid,
-    weigths_file=weights_dir / f"{mapluf_file.stem}_weights",
-)
 
 
 # %% change the categories
@@ -187,27 +182,22 @@ if USE_GNRF:
         # add this to the mapping
         ZH_2_GNFR |= splitted_cats
 
-    rasters_inv = group_categories(rasters_inv, ZH_2_GNFR)
+    inv = group_categories(inv, ZH_2_GNFR)
 
 # %% add the swiss inventory when needed
 if INCLUDE_SWISS_OUTSIDE:
     inv_ch = SwissRasters(
-        data_path=swiss_data_path,
-        rasters_dir=swiss_data_path.parent / "ekat_gridascii",
-        rasters_str_dir=swiss_data_path.parent / "ekat_str_gridascii_v17",
[…]
