### 2024-10-24 6d2a648
fixing brackets around some references and cleaning text
