+++ joss/paper.md
@@ -51,31 +51,34 @@ The simulations are often compared with measurements to verify if the declared e
 Figure \ref{fig:tno} presents an example of a gridded inventory.
 
 
-![CO2 emissions of the year 2015 from the TNOGHGco2_v6 inventory produced by TNO \label{fig:tno}.](raster_total_CO2.png){ width=100% }
+![CO2 emissions of the year 2015 from the inventory produced by TNO \label{fig:tno}.](raster_total_CO2.png){ width=100% }
 
 Inventories are created in multiple different formats and resolutions, which makes it difficult to compare and use them in atmospheric transport models.
 For example, TNO (Dutch Organization for Applied Scientific Research) provides inventories which contains both, area emissions on a regular grid and point
 sources at their exact locations.
 Other inventories, such as one from the city of Zurich, are provided as 
 vector data with various shapes (points, lines, polygons) depending on the category of the source.
-Atmospheric models require emission inventories to be in a specific file format. The atmospheric chemistry transport model ICON-ART [@icon_art_2.1], for example, requires emissions on a semi-structured triangular grid. As input for OEM it also requires temporal profiles to scale the emissions with hourly, daily and monthly variability.
+Atmospheric models require emission inventories to be in a specific file format.
+For example ICON-ART [@icon_art_2.1] requires emissions on a semi-structured triangular grid
+and temporal profiles to scale the emissions with hourly, daily and monthly variability.
 
 # Statement of need
 
 `emiproc` is a Python package that provides tools for processing emission inventories, harmonizing datasets, and preparing such data as inputs for atmospheric transport models.
-`emiproc` is designed to be flexible and extendable, allowing users to easily add custom functionality, to read new inventories or export data to other formats.
+Designed to be flexible and extendable, `emiproc` allows users to easily add custom functionality, read new inventories or export data to other formats.
 
 
-When modellers design their simulations, they are often interested in modifying the inventories. For example, they could do the following: scale the emissions based on future scenarios, aggregate emissions by sector or pollutant to simplify their simulations or combine multiple inventories to represent different sources such as anthropogenic and natural emissions.
+When modellers design their simulations, they are often interested in modifying the inventories.
+For example, they could do the following: scale emissions based on future scenarios, aggregate emissions by sector or pollutant to simplify their simulations or combine multiple inventories to represent different sources such as anthropogenic and natural emissions.
 `emiproc` provides this functionality and has already been successfully applied for different use cases:
 
-* [@acp-24-2759-2024] produced emission files for ICON-ART-OEM based on the EDGARv6
+* @acp-24-2759-2024 produced emission files for ICON-ART-OEM based on the EDGARv6
 inventory [@edgar_v6] (Emissions Database for Global Atmospheric Research).
 
-* [@donmez2024urban] conducted urban climate simulation using emissions produced 
+* @donmez2024urban conducted urban climate simulation using emissions produced 
 with `emiproc` for cities of Zurich and Basel.
 
-* [@ponomarev2024estimation] used `emiproc` to nest the Zurich city inventory
+* @ponomarev2024estimation used `emiproc` to nest the Zurich city inventory
 inside the Swiss national inventory and to further nest the Swiss inventory
 inside the European TNO inventory.
 
@@ -85,7 +88,8 @@ Compared to `HERMESv3`, which relies on specific configuration files, `emiproc` 
 
 # History
 
-An earlier version of `emiproc` was already published [@gmd-13-2379-2020], but it was limited to specific models and inventories. Starting in 2022 `emiproc` has been refactored to satisfy the requirements of high flexibility and modularity.
[…]
