### 2024-10-24 #7337 @lionel42
> An an example, line 27 has the reference in parentheses but it shouldn't be.
>
> As per the new comments, the paper length is now 1133. If you can shorten a little more that would be great, but if not, it is close enough.

I removed the parenthesis around citations where I understood they were not required.

Also I removed a bit more words. We are still above 1000 but I don't see where I could reduce more. I guess we can decide later based on the reviewer's comments.
