### 2024-11-06 #7337 @lionel42
@kthyng  What would be the next steps for the review ? Is it to contact reviewers ? https://github.com/openjournals/joss-reviews/issues/7337#issuecomment-2435229833
