### 2024-11-18 e74fa9d
Round all the dxs and dys to the number of decimal places given by the parameter "rounding"
