+++ docs/source/inventories.rst
@@ -58,6 +58,11 @@ Saunois
 
 :py:class:`emiproc.inventories.saunois.SaunoisInventory`
 
+CAMS_REG_AQ
+^^^^^^^^^^^
+
+:py:class:`emiproc.inventories.cams_reg_aq.CAMS_REG_AQ`
+
 
 
 Grids 
