+++ docs/source/inventories.rst
@@ -58,8 +58,8 @@ Saunois
 
 :py:class:`emiproc.inventories.saunois.SaunoisInventory`
 
-CAMS_REG_AQ
-^^^^^^^^^^^
+CAMS Regional Air Quality
+^^^^^^^^^^^^^^^^^^^^^^^^^
 
 :py:class:`emiproc.inventories.cams_reg_aq.CAMS_REG_AQ`
 
