### 2024-11-18 5507b99
Rounding dxs, dys vectorizes
