+++ emiproc/inventories/cams_reg_aq.py
@@ -13,6 +13,16 @@ from emiproc.profiles.vertical_profiles import read_vertical_profiles
 UNIT_CONVERSION_FACTOR = 1e9  # Tg -> kg
 
 class CAMS_REG_AQ(Inventory):
+    """The CAMS regional air quality inventory.
+
+    Contains gridded data of air pollutants (NOx, CO, CH4, VOC, NH3, SO2, PM2.5, PM10)
+    from the Copernicus Atmosphere Monitoring Service (CAMS).
+
+    You can access the data at
+    `CAMS-REG-ANT v6.1-Ref2
+    <https://eccad.sedoo.fr/#/metadata/608>`_
+
+    """
 
     grid: RegularGrid
 
@@ -56,6 +66,21 @@ class CAMS_REG_AQ(Inventory):
             "pm10":"PM10"
         },
     ):
+        """Create a CAMS_REG_ANT-inventory.
+
+        :arg nc_dir: The directory containing the NetCDF emission datasets. One file
+            per air pollutant.
+        :arg profiles_dir: The directory where the vertical and temporal profiles 
+            are stored. If None the directory nc_dir is used.
+        :arg year: Year of the inventory.
+        :arg substances_mapping: How to map the names of air pollutants from the 
+            names of the NetCDF files to names for emiproc. 
+        :arg categories_mapping: How to map the names of the emission categories from
+            the NetCDF files to names for emiproc.
+        :arg substances_mapping_profiles: How to map the names of air pollutants from
+            the vertical and/or temporal profiles to names for emiproc. Make sure this 
+            mapping is consistent with the substances_mapping.
+        """
         
         super().__init__()
 
