### 2024-11-18 15291c3
Restructured so that files selection is only done at one place.
