### 2024-11-18 #7337 @mengqi-z
👋 @jcla490 @dostuffthatmatters @einaraz - Would any of you be available to review this submission for JOSS? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
