+++ emiproc/inventories/cams_reg_aq.py
@@ -12,6 +12,7 @@ from emiproc.profiles.vertical_profiles import read_vertical_profiles
 
 UNIT_CONVERSION_FACTOR = 1e9  # Tg -> kg
 
+
 class CAMS_REG_AQ(Inventory):
     """The CAMS regional air quality inventory.
 
@@ -39,7 +40,7 @@ class CAMS_REG_AQ(Inventory):
             "sox": "SO2",
             "nh3": "NH3",
             "pm2_5": "PM25",
-            "pm10":"PM10"
+            "pm10": "PM10",
         },
         categories_mapping: dict[str, str] = {
             "A_PublicPower": "A",
@@ -53,7 +54,7 @@ class CAMS_REG_AQ(Inventory):
             "I_OffRoad": "I",
             "J_Waste": "J",
             "K_AgriLivestock": "K",
-            "L_AgriOther": "L"
+            "L_AgriOther": "L",
         },
         substances_mapping_profiles: dict[str, str] = {
             "nox": "NOx",
@@ -63,44 +64,42 @@ class CAMS_REG_AQ(Inventory):
             "so2": "SO2",
             "nh3": "NH3",
             "pm2_5": "PM25",
-            "pm10":"PM10"
+            "pm10": "PM10",
         },
     ):
         """Create a CAMS_REG_ANT-inventory.
 
         :arg nc_dir: The directory containing the NetCDF emission datasets. One file
             per air pollutant.
-        :arg profiles_dir: The directory where the vertical and temporal profiles 
+        :arg profiles_dir: The directory where the vertical and temporal profiles
             are stored. If None the directory nc_dir is used.
         :arg year: Year of the inventory.
-        :arg substances_mapping: How to map the names of air pollutants from the 
-            names of the NetCDF files to names for emiproc. 
+        :arg substances_mapping: How to map the names of air pollutants from the
+            names of the NetCDF files to names for emiproc.
         :arg categories_mapping: How to map the names of the emission categories from
             the NetCDF files to names for emiproc.
         :arg substances_mapping_profiles: How to map the names of air pollutants from
-            the vertical and/or temporal profiles to names for emiproc. Make sure this 
+            the vertical and/or temporal profiles to names for emiproc. Make sure this
             mapping is consistent with the substances_mapping.
         """
-        
+
         super().__init__()
 
-        filename_pattern = fr"CAMS-REG-ANT_EUR_0\.05x0\.1_anthro_(?P<substance>\w+)_v6\.1-Ref2_yearly_{year}\.nc"
+        filename_pattern = rf"CAMS-REG-ANT_EUR_0\.05x0\.1_anthro_(?P<substance>\w+)_v6\.1-Ref2_yearly_{year}\.nc"
 
         nc_dir = Path(nc_dir)
         if not nc_dir.is_dir():
-                raise FileNotFoundError(
-                    f"Profiles directory {nc_dir} is not a directory."
-                )
-        nc_files = [f for f in nc_dir.iterdir() 
-                    if f.is_file() 
-                    and f.suffix == '.nc'
-                    and re.match(filename_pattern, f.name)
-                    ]
-        
+            raise FileNotFoundError(f"Profiles directory {nc_dir} is not a directory.")
+        nc_files = [
+            f
+            for f in nc_dir.iterdir()
+            if f.is_file() and f.suffix == ".nc" and re.match(filename_pattern, f.name)
+        ]
+
         if not nc_files:
             raise FileNotFoundError(
                 f"No .nc files found matching the pattern '{filename_pattern}' in {nc_dir}"
-                )
+            )
 
         if profiles_dir is None:
             profiles_dir = Path(nc_dir)
@@ -110,7 +109,7 @@ class CAMS_REG_AQ(Inventory):
                 raise FileNotFoundError(
[…]
