### 2024-11-20 #7509 @mengqi-z
Hi @dostuffthatmatters,

Thank you for being transparent about potential conflicts of interest. After careful consideration, we believe there is a conflict of interest, particularly regarding the part "he is part of the scientific advisory board of [ITMS](https://www.fona.de/en/measures/funding-measures/ITMS_en.php) which my lab is partly funded by".

Unfortunately, this means you won’t be able to review this paper. But I really appreciate your willingness to contribute. I’m sure there will be other great opportunities for you to review JOSS papers in the future! Thank you again!
