### 2024-11-20 #7509 @mengqi-z
👋 @mikapfl @varsha2509 - Would any of you be interested in and able to take on this submission for JOSS?
