### 2024-11-21 1fc797d
CAMS Regional Air Quality Inventory  (#69)
* Round all the dxs and dys to the number of decimal
places given by the parameter "rounding"

* Updated link to download data

* CAMS regional air quality inventory v6.1-Ref2
added

* adding to doc

* improved the name

* Rounding dxs, dys vectorizes

* docstring added

* Restructured so that files selection is only done
at one place.

* permalink instead

* formatted black

* work with sub selection of substances

* remove profiles

* adding unit check

---------

Co-authored-by: coli <lionel.constantin@empa.ch>
