### 2024-12-02 #7509 @einaraz
Hi @lionel42! This is a great project, and I believe it will be valuable to many people. While this isn't my primary area of research, I tried evaluating the software itself rather than its specific application. Considering that this tool is likely to be used by individuals with different levels of Python expertise, I've suggested a few minor corrections and improvements for your consideration. I hope you find them helpful!

**1.** In the example EDGAR, maybe mention that the necessary folders must be created beforehand to save downloaded data.
**2.** In the example EDGAR, The script following the tutorial does not import matplotlib, which is required for the example to run.
**3.** When I ran the "Global Fire Emissions Database (GFED)" example, it required installing Dask, which I believe is used by xarray. The issue appeared when calling GFED5(). I ran the examples inside a virtual environment, where Dask was not previously available. Should this be added to the requirements list?
**4.** ICON-OEM Tutorial: In the "Preparing emission data for ICON-OEM" tutorial, could you add a link to download the TNO_GHGco_v4_0_year2018.nc file needed to run the model? If this file is proprietary, it would help to clarify its access conditions.
**5.** The tutorial currently shows how to process the EDGAR inventory. It might be helpful to include short examples for other inventories, as they seem to import different functions. Additionally, it would be useful to list which inventories can be downloaded with emiproc.
**6.** "Implementing a New Inventory" Section: This section may be unclear to users unfamiliar with Python or Object-Oriented Programming. Instead, I would suggest framing it as a guide on how to contribute to the code, so other users can add more functionalities.
**7.** Custom Grids: In the "Creating a custom grid" example, could you expand on how to use the new grid when analyzing inventories? A short example would make it more practical. Again, I'm thinking of users who might be less familiar with python.
**8.** I cloned the repo locally, and tried running pytest from the root directory (where the pyproject file is located). I got an error message related to the location of the file "heat.yaml". I then ran pytest outside of the root directory, and I passed that error (heat.yml found). However, it raises an additional error "FileNotFoundError: TNO Inventory file TNO_GHGco_v4_0_year2018.nc is not a file". Moreover, for the tests that did pass, plots showed up, which I don't think are very helpful to identify if the tests passed since I don't know how they should look like.
**9.** Overall, the tutorial is very detailed and contains enough information to get started. However, I noticed a few typos as well as some sentences that could be rephrased to sound more formal. I believe that a review of the text would improve readability and presentation.
**10.** As a small suggestion, I believe that moving the Tutorial section after you describe the available Inventories and Models would improve the flow.
**11.** Contributing: The documentation does not currently mention how to contribute to the code, only how to reach out for help. Maybe expanding on the section on adding new inventories could be a good start.

Please let me know if any of my comments are unclear!
Thank you!
