### 2024-12-04 e40c90c
adding corners to icon
