### 2024-12-04 bd4ea66
include irregular grids plots
