+++ docs/source/contrib/contribute.rst
@@ -1,0 +1,34 @@
+Contribute
+==========
+
+We are happy to receive contributions to emiproc.
+
+If you have questions or need help, you are welcome to open an issue on github.
+
+https://github.com/C2SM-RCM/emiproc/issues
+
+We are happy to answer your questions and help you to get started.
+
+
+Contribute as a developper
+--------------------------
+
+If you want to contribute to the code, the prefered way is via pull requests on github.
+
+https://github.com/C2SM-RCM/emiproc/pulls
+
+
+
+You can follow the tutorials below to understand how emiproc works and how to
+expand it.
+
+
+Tutorials for developpers
+-------------------------
+
+
+.. toctree::
+    :maxdepth: 2
+
+    new_inventory
+    create_grid
