+++ docs/source/tutos/tutorials.rst
@@ -33,6 +33,4 @@ The first tutorial to understand how emiproc works is the edgar_processing tutor
 
     edgar_processing
     gfed
-    new_inventory
     icon_oem
-    create_grid
