### 2024-12-04 7158f45
fixed pytest for contrib
