### 2024-12-04 7b026a1
add install and test to contrib
