### 2024-12-04 947fc1f
fix deprecation warning
