### 2024-12-04 b72d9a4
fix 'dims' deprecation warning
