+++ docs/source/tutos/tutorials.rst
@@ -1,9 +1,11 @@
 Tutorials
 =========
 
-Many examples of using emiproc can be found in the folder 
-examples from the github repository:
+If you want to start using emiproc, we recommend beginning with the tutorials below.
+They will guide you through the main features of emiproc.
 
+Additionally many examples of using emiproc can be found
+Additionally, many examples of using emiproc can be found in the GitHub repository.
 Examples 
 --------
 
@@ -16,16 +18,14 @@ Scripts
 -------
 
 Scripts were create on some dedicated cases. They are not well documented
-but there exist many for different cases. 
-
+Scripts were created for some dedicated cases. They are not well documented, but there are many scripts for different cases.
 https://github.com/C2SM-RCM/emiproc/tree/master/scripts
 
 
 Tutorials 
 ---------
 
-Tutorials provide guided examples with more in depth of how to use emiproc.
-
+Tutorials provide guided examples with more in-depth instructions on how to use emiproc.
 The first tutorial to understand how emiproc works is the edgar_processing tutorial.
 
 .. toctree::
