+++ docs/source/index.rst
@@ -3,10 +3,10 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-emiproc documentation
-=====================
+emiproc
+=======
 
-emiproc is a python package to generate emission input files for atmospheric 
+`emiproc` is a python package to generate emission input files for atmospheric 
 transport models.
 
 It can process inventories available in different formats and on different
@@ -19,7 +19,7 @@ models.
 A typical workflow is shown below:
 
 * Read an inventory
-* Perform some operations on it (e.g. regridding, merging categories, upsampling, ...)
+* Perform some operations on it (for example: regridding, groupping categories, upsampling, ...)
 * Export the inventory to a format that can be read by the model
 
 .. image:: diagrams/pipeline.drawio.svg
@@ -27,13 +27,13 @@ A typical workflow is shown below:
 Features 
 --------
 
-* Support of multiple inventories like EDGAR, TNO-CAMS
+* Support multiple inventories such as EDGAR and TNO-CAMS
 * Conservative spatial regridding
-* Exporting to different formats (icon-art, cosmo-art, netcdf rasters)
-* Spatially merging inventories around a region
+* Export to different formats (ICON-ART, COSMO-ART, NetCDF rasters)
+* Merge inventories spatially around a region
 * Separate handling of point sources if desired
-* Categories/Substance selection
-* Re-grouping of emission categories
+* Group emission categories
+* Speciation of emissions
 * Visualization of the output
 * Cyclic temporal profiles or time series of emissions
 
