### 2024-12-04 #7509 @lionel42
Hi @einaraz ,

Thanks a lot for the review.
It is very valuable to have all these comments and I tried to do my best to improve according to your suggestions.

I have implemented/fixed most of your points in a recent merge request (https://github.com/C2SM-RCM/emiproc/pull/70).

As you suggested I created a [Contributing section](https://emiproc.readthedocs.io/en/master/contrib/contribute.html) in the documentation about how to contribute, in which I included the tutorials which are more oriented towards developers than users (implementing new inventories or new grids).

I also added there documentation about how to run the tests. I think this should fix the issues you had when running the tests.

Regarding the TNO file in the ICON-OEM tutorial, we have the problem that the data is not freely available. I decided to instead replace it with the EDGAR inventory, so that anyone can run this tutorial.

Thanks again for the suggestions and please let me know if you have other ones.
Best,
Lionel
