### 2024-12-06 #7509 @einaraz
Hi @lionel42,

Thank you so much for addressing my comments. I think that the flow of the tutorial improved and is more linear. I also like
the expansion on how to contribute to the code.

I was able to run pytest successfully, although not from the root directory (there I still get an error related to the heat.yaml file location). Running pytest from one directory above the root directory worked.

Thank you!
Einara
