### 2024-12-06 #7509 @lionel42
@einaraz

Thanks a lot,

I had a look again to this `heat.yaml` file issue, and I really don't know what could be the problem.
If you manage to make it work by changing the directory, it is okay for me, otherwise you are welcome to open an issue about this.

Best,
Lionel
