### 2024-12-07 #7509 @einaraz
Thank you, @lionel42. Maybe the problem is just on my end.

@mengqi-z, I have no further comments on this project. I believe it adheres to the guidelines for acceptance at JOSS.

Best,
Einara
