+++ docs/source/tutos/tutorials.rst
@@ -34,3 +34,4 @@ The first tutorial to understand how emiproc works is the edgar_processing tutor
     edgar_processing
     gfed
     icon_oem
+    regridding
