### 2024-12-18 #7509 @mikapfl
Hi @lionel42 ,

Just a quick question: where do I find the code used to generate Figure 1 (`raster_total_CO2.png`)? I'd like to check if I can reproduce it.

Best,

Mika
