### 2024-12-18 3aa1924
Add ipython as docs dependency for syntax highlighting
