### 2024-12-18 efb73d7
Add cams_reg_aq.CAMS_REG_AQ to API docs
