### 2024-12-18 1883eff
Typos and formatting in the tutorial docs
