### 2024-12-18 bd42053
remove more duplicates in the docs
