### 2024-12-18 #7509 @lionel42
@mikapfl
You can find the script there: https://github.com/C2SM-RCM/emiproc/blob/master/scripts/plot_tno.py

However, to reproduce you will need the tno data which sadly is not publicly available.

I could instead put a plot of the edgar inventory, which is publicly available.
What do you think ?
