### 2024-12-18 #7509 @mikapfl
@lionel42 Given that I have to tick off "Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers." for the review, I would be more comfortable with a plot based on Edgar, yes. Arguably, the plot is not an original result, but rather an illustrative example, but if it is easy for you to include a reproducible result instead, that would be great. It looks like something from the edgar tutorial would be equally good and I was already able to reproduce it. (-:
