### 2024-12-18 aceb2b1
update the picture with edgar instead of tno
