### 2024-12-18 #7509 @lionel42
@mikapfl  Makes sense to me, I have replaced it with edgar in https://github.com/C2SM-RCM/emiproc/commit/aceb2b1ae63dac37a0e44b622f2f2c0fb00377cc

The new plot script is available at https://github.com/C2SM-RCM/emiproc/blob/coli/scripts/plots/plot_for_joss.py
