### 2024-12-18 #7509 @mengqi-z
@mikapfl - Wonderful! Thanks for your review and positive feedback on this submission!

@lionel42 - Both reviewers have completed their reviews. I’ll proofread the paper and ensure it meets JOSS standards, then follow up with you on the next steps. Thanks!
