### 2025-01-06 C2SM-RCM/emiproc#77 @mengqi-z
JOSS Review
Hi @lionel42,

Thank you for your submission  to JOSS https://github.com/openjournals/joss-reviews/issues/7509. I have few comments on the paper for you to address before we can move to the next phase.

1. The following references are missing co-author names. Please review all references and ensure that the full list of authors is included.
    * Ponomarev, S., N., & Brunner, D. 2024  https://doi.org/10.5194/egusphere-egu24-7420
    * Dönmez, E., K., & Brunner, D. 2024 https://doi.org/10.5194/egusphere-egu24-3375

2. Citation for "ICON partnership, 2024": According to the official suggested citation from this link https://www.wdc-climate.de/ui/entry?acronym=IconRelease01, "ICON partnership (DWD; MPI-M; DKRZ; KIT; C2SM)" should be treated as a single author entity, not as multiple authors.

3. There is an extra space after "API documentation" on line 66. Please remove it.

4. On line 82, the word "Edgar" should be capitalized.
