+++ joss/paper.bib
@@ -76,7 +76,7 @@ doi = {10.2800/795737}
       title = {ICON release 2024.01},
       publisher = {World Data Center for Climate (WDCC) at DKRZ},
       year = {2024},
-      author = {ICON partnership, and DWD and MPI-M and DKRZ and KIT and C2SM},
+      author = {{ICON partnership (DWD; MPI-M; DKRZ; KIT; C2SM)}},
       doi = {10.35089/WDCC/IconRelease01}
 }
 
@@ -126,7 +126,7 @@ DOI = {10.5194/acp-24-2759-2024}
 }
 
 @inproceedings{donmez2024urban,
-  author    = { Dönmez, K., Emmenegger, L., and Brunner, D.},
+  author    = { Dönmez, K. and Emmenegger, L. and Brunner, D.},
   title     = {Urban Climate and CO2 Simulations with the New Atmospheric Model ICON-ART Accounting for Spatially Varying Urban Morphology and Material Properties},
   booktitle = {EGU General Assembly 2024},
   year      = {2024},
@@ -138,7 +138,7 @@ DOI = {10.5194/acp-24-2759-2024}
 }
 
 @inproceedings{ponomarev2024estimation,
-  author    = {Ponomarev, N., Steiner, M., Koene, E., Constantin, L., Rubli, P., Grange, S., Emmenegger, L., and Brunner, D.},
+  author    = {Ponomarev, N. and Steiner, M. and Koene, E. and Constantin, L. and Rubli, P. and Grange, S. and Emmenegger, L. and Brunner, D.},
   title     = {Estimation of CO2 fluxes in the city of Zurich using the mesoscale atmospheric transport and inversion model ICON-ART-CTDAS},
   booktitle = {EGU General Assembly 2024},
   year      = {2024},
