### 2025-01-07 C2SM-RCM/emiproc#77 @lionel42
@mengqi-z  Thanks for reporting.

I have updated the joss branch with your suggested corrections.
