### 2025-01-08 #7509 @lionel42
Version: `v2.3.1`

https://github.com/C2SM-RCM/emiproc/releases/tag/v2.3.1
