### 2025-01-08 #7509 @lionel42
The doi on zenodo:

https://doi.org/10.5281/zenodo.14614230
