### 2025-01-08 #7509 @mengqi-z
@lionel42 - Thanks! I noticed a couple of discrepancies in the Zenodo archive:

- The order of authors in Zenodo doesn't match the paper. Could you please update the Zenodo author list to reflect the order in the paper?
- The version number in GitHub release is `v2.3.1`, but the Zenodo archive lists it as `v2.3-joss`. Could you update the Zenodo version to match the GitHub release (`v2.3.1`)?

Thanks!
