### 2025-01-09 #7509 @lionel42
@mengqi-z

Thanks for finding that out and sorry for the discrepancies.

I have now fixed authors order and version.
