### 2025-01-09 #7509 @mengqi-z
@lionel42 – I’ve recommended your submission for acceptance. Thanks so much for submitting to JOSS! From here, our Editor-in-Chief will step in to do some final checks before the paper can be officially accepted and published. Congrats and great job!

@einaraz @mikapfl – Huge thanks to all of you for reviewing this paper! It’s been great working with you during this process, and I really appreciate your thoughtful feedback. I look forward to seeing your reviews for JOSS in the future!
